# Simplify & optimize Replace