… src/LinkDotNet.StringBuilder/IntegerSpanList.cs
… src/LinkDotNet.StringBuilder/NaiveSearch.cs