There are two reasonable approaches to replacing a value with another value of a different length:
1. Find every index of oldValue, then create a bigger/smaller buffer once and insert the values.
2. Find each index of oldValue, replacing with newValue and resizing the buffer each time.

Previously, the replace operation did a combination of the two which seemed over-the-top. I changed it to only use the second approach, which is far simpler and gives better performance!

**BEFORE:**

| Method              | Mean      | Error    | StdDev   | Ratio | RatioSD | Gen0   | Allocated | Alloc Ratio |
|-------------------- |----------:|---------:|---------:|------:|--------:|-------:|----------:|------------:|
| DotNetStringBuilder |  18.66 us | 0.266 us | 0.249 us |  1.00 |    0.02 | 9.4604 |  29.08 KB |        1.00 |
| ValueStringBuilder  | 122.50 us | 0.732 us | 0.685 us |  6.57 |    0.09 | 4.8828 |  15.24 KB |        0.52 |

**AFTER:**

| Method              | Mean     | Error    | StdDev   | Ratio | Gen0   | Allocated | Alloc Ratio |
|-------------------- |---------:|---------:|---------:|------:|-------:|----------:|------------:|
| DotNetStringBuilder | 18.34 us | 0.177 us | 0.166 us |  1.00 | 9.4604 |  29.08 KB |        1.00 |
| ValueStringBuilder  | 19.39 us | 0.184 us | 0.163 us |  1.06 | 4.3945 |  13.55 KB |        0.47 |

**Note:**
Two of the tests had to be changed because, in my opinion, you shouldn't be allowed to give out-of-bounds parameters just because another parameter is empty and short-circuits.