@Joy-less 2025-02-21
I also added `scoped` to the `ReplaceGeneric` `oldValue` argument. I noticed that `ReplaceGeneric` can simply be renamed to `Replace`, would you like this? @linkdotnet