@linkdotnet 2025-02-21
> I also added `scoped` to the `ReplaceGeneric` `oldValue` argument. I noticed that `ReplaceGeneric` can simply be renamed to `Replace`, would you like this? @linkdotnet

Really? I thought it might resort to a boxed invocation. We can also do this for v3 if that isn’t the case. For now, because of semantic versioning, I would like to keep it as is. The scoped though is absolutely fine :)