@Joy-less 2025-02-21
I made some minor clarity changes. It's ready to be merged unless you have more concerns.